requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2.3
git+https://github.com/kemalayhan/django-rest-swagger.git
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
idna==3.4
ipdb==0.13.13
ipython==8.12.2
itypes==1.2.0
jedi==0.19.0
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
openapi-codec==1.3.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
